refactor http request
