--- lazyllm/components/http_request/__init__.py
@@ -1,5 +1,0 @@
-from .http_request_node import HttpRequestNode
-
-__all__ = [
-    "HttpRequestNode"
-]
