--- lazyllm/engine/status.py
@@ -1,7 +1,0 @@
-class NodeExecutionStatus():
-    """
-    Workflow Node Execution Status Enum
-    """
-    RUNNING = 'running'
-    SUCCEEDED = 'succeeded'
-    FAILED = 'failed'
