+++ lazyllm/tools/http_request/__init__.py
@@ -1,0 +1,5 @@
+from .http_request import HttpRequest
+
+__all__ = [
+    "HttpRequest"
+]
